pyproject.toml
[project]
name = "youtube-agent"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "langchain==1.1.2",
    "langchain-community>=0.3.2",
    "langchain-openai",
    "langchain-ollama",
    "streamlit==1.40.0",
    "pyyaml==6.0.2",
    "duckduckgo-search==6.1.0",
    "ddgs",
    "youtube-transcript-api>=0.6.3",
    "pydantic>=2.7.4,<3.0.0",
    "pytube>=15.0.0"

]

[tool.uv]
index-url = "https://pypi.org/simple"